requirements.txt
yfinance
pandas
numpy
scipy
scikit-learn
giotto-tda
stable-baselines3
torch
gymnasium
statsmodels
arch
newsapi-python
pandas-datareader
fredapi
requests
schedule
duckdb
loguru
tqdm
pyzmq
joblib
optuna
matplotlib
python-dotenv
vaderSentiment
